requirements.txt
pyspark==3.5.5
flask
psycopg2-binary
flask-cors
pyyaml